Feat: order_form에서 해당상품에 투자시, 진행되는 작업들에 대하여 Transaction 처리

+++ src/main/java/com/fundfun/fundfund/dto/order/InvestDto.java
@@ -9,5 +9,5 @@ import javax.validation.constraints.NotNull;
 @Setter
 public class InvestDto {
     @NotNull(message = "투자금액을 입력해주세요.")
-    private Integer cost;
+    private Long cost;
 }
